# openapi/auth.yaml
openapi: 3.0.3
info:
  title: Growfoundry Authentication API
  version: 2.0.0
  description: Authentication endpoints with separated auth and profile tables
paths:
...
  /api/auth/sessions:
    post:
      summary: User login
      description: Authenticates user and returns access token. For web clients, sets httpOnly refresh token cookie. For mobile/desktop/server clients, returns refreshToken in response body.
      tags:
        - Client
      parameters:
        - name: client_type
          in: query
          schema:
            type: string
            enum: [web, mobile, desktop, server]
            default: web
          description: |
            Client type determines how refresh tokens are returned:
            - web: Refresh token stored in httpOnly cookie, csrfToken returned in response
            - mobile/desktop/server: refreshToken returned directly in response body
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserResponse'
                  accessToken:
                    type: string
                  csrfToken:
                    type: string
                    nullable: true
                    description: CSRF token for use with refresh endpoint (web clients only)
                  refreshToken:
                    type: string
                    nullable: true
                    description: Refresh token for mobile/desktop/server clients (null for web clients)
        '401':
          description: Invalid credentials
        '403':
          description: Email verification required
  /api/auth/refresh:
    post:
      summary: Refresh access token
      description: |
        Refresh access token using refresh token.
        - Web clients: Use httpOnly refresh token cookie with X-CSRF-Token header
        - Mobile/Desktop/Server clients: Send refreshToken in request body
      tags:
        - Client
      parameters:
        - name: client_type
          in: query
          schema:
            type: string
            enum: [web, mobile, desktop, server]
            default: web
          description: |
            Client type determines how refresh tokens are handled:
            - web: Refresh token from httpOnly cookie, requires X-CSRF-Token header
            - mobile/desktop/server: refreshToken provided in request body, new refreshToken returned in response
        - name: X-CSRF-Token
          in: header
          schema:
            type: string
          description: CSRF token received from login/register response (required for web clients only)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: Refresh token (required for mobile/desktop/server clients only)
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserResponse'
                  accessToken:
                    type: string
                  csrfToken:
                    type: string
                    nullable: true
                    description: New CSRF token for subsequent refresh requests (web clients only)
                  refreshToken:
                    type: string
                    nullable: true
                    description: New refresh token for mobile/desktop/server clients (must be persisted for next refresh)
        '401':
          description: No refresh token provided or user not found
        '403':
          description: Invalid CSRF token
...
  /api/auth/sessions/current:
    get:
      summary: Get current session
      description: |
        Returns the currently authenticated user's basic info from the access token.
        Project admin tokens return a projectAdmin session object instead of an auth.users row.

        This endpoint does not refresh expired access tokens by itself.
        For browser apps using the TypeScript SDK, call `auth.getCurrentUser()`
        during startup. The SDK will use the httpOnly refresh cookie automatically
        when it can refresh the session. Server, mobile, and other non-browser
        clients should call `/api/auth/refresh` explicitly.
      tags:
        - Client
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current session info
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    required:
                      - user
                    additionalProperties: false
                    properties:
                      user:
                        $ref: '#/components/schemas/UserResponse'
                  - type: object
                    required:
                      - projectAdmin
                    additionalProperties: false
                    properties:
                      projectAdmin:
                        $ref: '#/components/schemas/ProjectAdminResponse'
        '401':
          description: Unauthorized
  /api/auth/admin/sessions:
    post:
      summary: Admin login
      description: Authenticates admin user for dashboard access
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
      responses:
        '200':
          description: Admin login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  projectAdmin:
                    $ref: '#/components/schemas/ProjectAdminResponse'
                  accessToken:
                    type: string
                  csrfToken:
                    type: string
                    description: CSRF token for `/api/auth/admin/refresh`
        '401':
          description: Invalid credentials
        '403':
          description: User is not an admin
  /api/auth/admin/sessions/exchange:
    post:
      summary: Exchange cloud provider authorization code for admin session
      description: Verifies an authorization code/JWT from from Growfoundry Cloud platform and issues an internal admin session token with project_admin role
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
                  description: Authorization code or JWT from the Growfoundry
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Cloud authorization verified, admin session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  projectAdmin:
                    $ref: '#/components/schemas/ProjectAdminResponse'
                  accessToken:
                    type: string
                    description: Internal JWT for admin authentication
                  csrfToken:
                    type: string
                    description: CSRF token for `/api/auth/admin/refresh`
        '400':
          description: Invalid authorization code or JWT verification failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/admin/refresh:
    post:
      summary: Refresh admin dashboard access token
      description: Uses the dashboard-only httpOnly `growfoundry_admin_refresh_token` cookie with an `X-CSRF-Token` header.
      tags:
        - Admin
      parameters:
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
          description: CSRF token returned from admin login or admin refresh
      responses:
        '200':
          description: Admin token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  projectAdmin:
                    $ref: '#/components/schemas/ProjectAdminResponse'
                  accessToken:
                    type: string
                  csrfToken:
                    type: string
                    description: CSRF token for the next admin refresh request
        '401':
          description: No admin refresh token provided or project admin not found
        '403':
          description: Invalid CSRF token
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    ProjectAdminResponse:
      type: object
      required:
        - subject
        - username
      properties:
        subject:
          type: string
          example: local:admin
        username:
          type: string
          example: admin
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        profile:
          type: object
          nullable: true
          additionalProperties: true
          description: User profile data (name, avatar_url, and custom fields)
          properties:
            name:
              type: string
            avatar_url:
              type: string
              format: uri
        metadata:
          type: object
          nullable: true
          additionalProperties: true
          description: System metadata (device ID, login IP, etc.)
        emailVerified:
          type: boolean
        providers:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - statusCode
      properties:
        error:
          type: string
          description: Error code for programmatic handling
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Email is already in use"
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
        nextActions:
          type: string
          description: Suggested action to resolve the error
          example: "Please use a different email address"